requirements.txt
requests==2.28.2
allure-pytest==2.13.2
pytest==7.4.4